# fusecpt-backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Node.js + TypeScript backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/super-admin/users": {
      "post": {
        "summary": "Create a new user (Admin or User)",
        "tags": ["SuperAdmin"],
        "description": "Only accessible by logged-in super-admin. Requires `manage:all` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "role"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "user"],
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "Required only if role is \"user\". Ignored for \"admin\".",
                    "example": "userPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error / missing fields"
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "403": {
            "description": "Forbidden (not super-admin)"
          }
        }
      }
    }
  },
...
}